requirements.txt
beautifulsoup4>=2.0.1
flask>=1.1.2
flask_marshmallow>=0.14.0
flask_sqlalchemy>=1.3.20
marshmallow_enum>=1.5.1
marshmallow-sqlalchemy>=0.24.1
pillow>=8.1.0
python-slugify>=4.0.1
requests>=2.24.0
werkzeug==0.16.1
